Replaced files with most up to date ones.
Minor bug fixes and parameter prior alterations. Everything works well at this point, just need to sort out continuum 'baseline' and tone the priors for P-Cygni profiles slightly.
Too many P-Cygni profiles are determined when the true profile is an absorption component. See if the baseline correction in the next version changes this or if the profile-determination method has to be more stringent..